+++ NEWS.md
@@ -1,3 +1,5 @@
+# pins (development version)
+
 # pins 1.2.0
 
 ## Breaking changes