+++ NEWS.md
@@ -1,6 +1,6 @@
 # pins (development version)
 
-* New environment variable `PINS_CACHE_PATH` controls the location of the 
+* New environment variable `PINS_CACHE_DIR` controls the location of the 
   default cache path (#748).
 
 # pins 1.2.0